+++ valid-QRDA-III-latest.qpp.json
@@ -1,0 +1,96 @@
+{
+  "performanceYear" : 2017,
+  "entityType" : "individual",
+  "taxpayerIdentificationNumber" : "000777777",
+  "nationalProviderIdentifier" : "0777777777",
+  "measurementSets" : [ {
+    "category" : "quality",
+    "submissionMethod" : "electronicHealthRecord",
+    "measurements" : [ {
+      "measureId" : "236",
+      "value" : {
+        "isEndToEndReported" : true,
+        "performanceMet" : 800,
+        "eligiblePopulationExclusion" : 50,
+        "performanceNotMet" : 150,
+        "eligiblePopulation" : 1000
+      }
+    }, {
+      "measureId" : "001",
+      "value" : {
+        "isEndToEndReported" : true,
+        "performanceMet" : 900,
+        "performanceNotMet" : 50,
+        "eligiblePopulation" : 950
+      }
+    }, {
+      "measureId" : "130",
+      "value" : {
+        "isEndToEndReported" : true,
+        "performanceMet" : 800,
+        "eligiblePopulationException" : 20,
+        "performanceNotMet" : 180,
+        "eligiblePopulation" : 1000
+      }
+    }, {
+      "measureId" : "371",
+      "value" : {
+        "isEndToEndReported" : true,
+        "strata" : [ {
+          "performanceMet" : 486,
+          "eligiblePopulationExclusion" : 35,
+          "performanceNotMet" : 79,
+          "eligiblePopulation" : 600,
+          "stratum" : "4MonthsOfEnd"
+        }, {
+          "performanceMet" : 700,
+          "eligiblePopulationExclusion" : 40,
+          "performanceNotMet" : 60,
+          "eligiblePopulation" : 800,
+          "stratum" : "4&8MonthsAfterStart"
+        }, {
+          "performanceMet" : 520,
+          "eligiblePopulationExclusion" : 40,
+          "performanceNotMet" : 20,
+          "eligiblePopulation" : 580,
+          "stratum" : "4MonthsOfStart"
+        } ]
+      }
+    } ],
+    "programName" : "mips",
+    "performanceStart" : "2017-01-01",
+    "performanceEnd" : "2017-12-31"
+  }, {
+    "category" : "aci",
+    "submissionMethod" : "electronicHealthRecord",
+    "measurements" : [ {
+      "measureId" : "ACI_PEA_1",
+      "value" : {
+        "numerator" : 600,
+        "denominator" : 800
+      }
+    }, {
+      "measureId" : "ACI_EP_1",
+      "value" : {
+        "numerator" : 400,
+        "denominator" : 600
+      }
+    } ],
+    "programName" : "mips",
+    "performanceStart" : "2017-02-01",
+    "performanceEnd" : "2017-05-31"
+  }, {
+    "category" : "ia",
+    "submissionMethod" : "electronicHealthRecord",
+    "measurements" : [ {
+      "measureId" : "IA_EPA_3",
+      "value" : true
+    }, {
+      "measureId" : "IA_CC_10",
+      "value" : true
+    } ],
+    "programName" : "mips",
+    "performanceStart" : "2017-01-01",
+    "performanceEnd" : "2017-04-30"
+  } ]
+}
